Share replay and runner details in details dir
